First demo for Prime G1
This is a very limited demo firmware for Prime G1 which can be flashed
to the calc. It does nothing more than light the 3 top LEDs.

A fixed header has to be added to the binary and file size has to be
corrected to the size stated in the header as follows:

cat tools-bin/p1_header tools-bin/newrplp1.bin >PRIME_OS.ROM
truncate --size=1M PRIME_OS.ROM

Then PRIME_OS.ROM can be replaced for flashing in recovery mode v13.