Improved error handling in NAND code.
Functions now propagate meaningful error codes.
Removed exceptions from NAND code, especially a misleading one and moved
them to SDDriver.
Error codes are not propagated to the user though.
Contains an additional wait in block erase code which is inspired by PRIME
firmware, but this might settle the same thing as the improved NANDCheckWrite
code of the last commit.