Build different makefiles depending on the config
This is to avoid getting the wrong options when building something.
Notice that this does not force the rebuild of the final firmware,
so if you perform a debug and a release build back to back, and then
do another debug build, it may not re-link, since there is already a binary
there and no need to rebuild it.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>