Generate proper UNIFONT structures for fonts
For type safety and efficiency of the generated code, it is better if we genrate
a proper UNIFONT data structure for the fonts. Also, this cleans up the code
using fonts, which no longer needs ugly and somewhat dangerous (UNIFONT *) casts.

Having a proper UNIFONT structure and build rules will make it easier to add
fonts later if we want.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>