Drop one level of indirection for font from objects
Everything in the code was using one extra level of indirection to get the
UNIFONT pointer from the object. I could not find a valid reason for that extra
level of indirection. The only rationale I could think of was if a variable
contains a FONT object and its value is updated. However, this should be
infrequent enough that we could update the halScreen.FontArray when we detect
that case by calling halUpdateFontArray.

There seems to be some code trying to do that and caching information using a
font hash, so I suspect this is the intent.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>