Add rule for clang-format application
This rule is parallelized for efficiency.

The Microsoft style was chosen after quick experiments showing that it
was the one causing the smallest number of diffs according to git
diff-stat. Tested with both all the files at top-level (the only ones
with C++ in it) and with all the files matchin `newrpl/a*.c`. That
should be a good enough sample of the overall style.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>