One more step towards the transition to strong typing in ggl
Wrote a new implementation of ggl_op that, at least in theory, can replace all
the existing primitives. Inlining should eliminate the parts of the code that
are not useful in a specific context. This was tested on simple, standalone
examples, but it remains to be validated that the compiler inlining budget is
sufficient at -O3 to get good code.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>