Connect bitmap cache to ggl_copy correctly
We now pass arguments as function parameters rather than inside the gglsurface
strucutre, the objective being to have gglsurface become as "static" as
possible (it's a rather inefficient way to pass parameters around).

Also changed the order of parameters in uiDrawBitmap to follow the convention
that the target comes before the source.

This version still garbles the output. This is probably related to the
computation of offsets for small bitmaps.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>