Remove all the filtering code
Prepared some (currently unused) ggl filters for darkening, etc. But first need
to fix a few other issues, and decide whether we really care about this specific
visual effect. If we do, there is a better way to do it in one pass instead of
successive calls to the filter function.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>