Major refactoring of keyboard code
This is a week-end long effort to simplify and improve the keyboard code.
The overall goal is to make the shift-handling isolated and target-dependent,
and to pave the way for faster/simpler per-plane lookup of keyboard functions.
There were major simplifications in the handling of long presses and held keys
as well.

The new code also paves the way for system and setup functions controlled, in
the old HP tradition, by holding the ON key and another key. For example, ON+
and ON- can be used to adjust the contrast, ON. to adjust the locale, and so
on. This is not connected yet, but took over the old exception code for ON-A-C
and ON-A-F (knowing that I fully expect ON by itself to interrupt a program,
without having to resort to ON-A-C, but this is for later).

In the current state, the refactored code correctly handles the shift logic for
all three platforms. The logic is adapted to the physical keyboard layout:
- On the HP50G and related, there are three shift functions, and they operate
  relatively independently, except that Shift-Alpha changes the lower-case
  state (which at some point will need its annunciator).
- On the Prime, there is a Shift and an Alpha key. The Shift key can now be
  pressed twice or held to get a right shift.
- On the DM42, there is a single Shift, no Alpha, so the Shift is a three-way
  toggle Shift -> Alpha -> None. A long-press of Shift gets you a right shift.

The keyboard does not work fully yet, notably in alpha mode, because in the
process, I discovered how hacky the old code was in that respect, at least on
Prime. Since the Prime (and the DM42) have some alphabetic characters on the
numeric keys, unlike the HP50G, the old code was remapping these to keys that
did not exist on the HP50G, e.g HOME. That made it possible to get the alpha
mode to "work", but required extra keys. That solution can't work on the DM42,
which has less keys. A proper handling of Alpha mode is required, which does
not exist at all in the current code, and will need ot be created.

Here is the detail of some of the changes made:
- Move shared declarations to new file, keyboard.h
- Reorganize keyb_irq_getmatrix() to share the debouncing code.
- Move keyboard declarations to keyboard.h, taking them away from hal_api.h
- Organized various keyboard layers, with grouping and naming conventions
- Move GPFDAT to the HP50G target.h header, use HWREG() macro.
- Many comments added or changed
- Various code reordering / code moves
- A lot of global renaming to make style more consistent and readable
- Move a lot of the shared keyboard logic to keybcommon.c
- Rework matrix scanning loops on Prime and 50G to use a single column loop
- clang-format a lot of the code I was working on
- Removed a lot of the doxygen documentation for internal stuff
- Remove a number of unused macros, e.g. keyb_isON()
- Add a type for keyboard messages, keyb_msg_t (useful for LLDB formatting)
- Make keyboard buffer entries 32-bit on all platforms
- Change keyb_set_shift_plane to take a bitmap instread of 4 ints

Despite adding support for a brand-new per-target shift logic, the new code is
roughly 800 lines shorter.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>