Switch to pattern-based palette, and simplify menu system
The last step towards a fully pattern-based color palette.
This simplifies the code quite a bit, getting rid of the last remnants of the
"old" GGL.

In the process, reworked the menu display quite significantly so that the

As a result, most UI elements now show up correctly on the DM42 simulator,
including nice patterned line separators for the command line. The values for
the default theme are chosen to avoid introducing too much noise in the
black-and-white UI (i.e. no pattern for the command line background) while at
the same time reinforcing separate UI elements on sytems capable of gray-scale.

During this refactoring, I noticed that there were already handlers for ON+ and
ON- to adjust the contrast. I will need to connect them to the new "attention"
code introduced by the recent keyboard refactoring.

Another 300 lines of code removed overall.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>