Refactor and simplify key handling code
Instead of a linear search through a hard-to-maintain table with a complex
context management logic, simply lookup by keycode in arrays, with a shared
logic that indicates which key plane we consider at any given time.

This vastly simplifies the overall keyboard-handling logic, and makes it
possible to easily implelent vastly different keymaps and key logic for
different platforms. Notably, the DM42 has a completely different keyboard
style, with only two arrow keys. It is important that the mapping of keys to
functions be implemented in easily configurable tables, rather than through
logic that is dispersed in the various handlers and includes 50G assumptions.

The work is not finished. A lot of the handlers will have to be reworked to make
them less 50G-dependent. Also, many new fake menus were created, which will need
ot be adjusted to match the actual menu location (or created if they don't
exist).

The intent long-term is to make sure that each platform still works with its own
logic, i.e. the Prime takes advantage of its application keys and the DM42 of
its dedicated functions keys, rather than try to impose a 50G keyboard logic on
platorms where it won't work well.

With this commit, the DM42 simulator is practically usable.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>