Improve handling of long-press
This is far from perfect, however. It looks like only the terminal menu
handlers (i.e. not directories) know how to deal with long-press to display
help. This needs to be changed if we want to properly implement a "help on key
press" that shows the name of the function on press and then help after a long
press.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>