Add rule to build documentation extractor
This extracts the built-in help from the commands and turns it into a series of
markdown files for GitHub/GitLab.

Ultimately, a help reader in newRPL will be able to take advantage of the
generated markdown files, instead of relying on in-ROM messages. This will
facilitate translation and customization of the help files, much like on the
DM42.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>