Refactor help engine
The new help system works at two level:

- On key down, show the name of the command being used
- On long press, show extended help for the command

The extended help for now is not implemented yet, though some preliminary work
was done in the 'docs' branch for a tool that extracts the built-in help into a
file. The idea is to have a built-in help file in markdown format, with the
possiblity to load an updated file from disk, e.g. with your own commands, or
localized in some other language.

In the process, some significant changes in all string-manipulation routines to
enforce the convention that read-only text pointers have utf8_p type, and
read-write have byte_p type. Eliminated a lot of useless casts in the process.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>