Draw help on the whole screen
Since the intent is to have much more extended help, it's better to draw it on
the whole screen. At some point, we'll need a way to scroll through the help if
it's more than a screenfull, which is more likely to happen on the HP50g because
of the smaller screen.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>