Display help from a built-in help file
The built-in help file is built by concatenating the various MD files.
This makes it possible to hand-edit the help, and also ensures that whatever
shows on the web site matches the integrated built-in help.

At some point in the future, we will want to be able to load a help file from
the SD card or built-in memory, to allow for customization.

Also, do we want navigation? That seems like a lot of work, not sure it really
adds compared to what we have now.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>