Layout rendering engine
In order to have more flexibility for how to render elements on screen,
implement a layout rendering engine that makes it possible to describe the
logical layout of the screen for each platform.

Despite the addition of new files and a whole new high-level layout engine, this
removes about 375 lines of code.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>