Insert lowercase or uppercase characters based on key message
The keyboard logic records the state of alpha keys at the time they are pressed,
so it is more correct to select upper/lowercase based on what is in the keyboard
message (buffered) rather than what is in the cursor state (not buffered).

This restores full keyboard functionality as far as I can tell.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>